requirements.txt
streamlit>=1.30.0
pdfplumber>=0.10.0
python-docx>=1.1.0
spacy>=3.7.0
pytesseract>=0.3.10
pdf2image>=1.16.0
openai>=1.12.0
python-dotenv>=1.0.0